Also allow python imports if the first line is shebang
Previously importing python modules didn't work because the function is
only run if the file type is .py. This change will also run the add
import function if the first line in the file is the python shebang

This is important because often scripts I put in my home bin directory
dont' have the py extension.