Add Ctags generation
Add a shortcut to generate a ctags file from all the python files recursively
in the current directory.

To do this we needed a new python function that can call an external command
and pass a string to the external command's standard input.

also made <leader>b open/close the buffer menu. The default is to just open
it and do nothing if you hit it again.